requirements.txt
paste<2.0
WebOb<1.4
alembic>=0.4.2
venusian
pydns
lxml
PIL
psycopg2
pyramid
pyramid_debugtoolbar
pyramid_tm
SQLAlchemy>=0.8.0b2
transaction
waitress
zope.sqlalchemy
reportlab

# Authentication transition
repoze.who>2.0
repoze.who.plugins.sa
repoze.who-friendlyform
pyramid_who

#### LEGACY (zkpylon) requirements:
AuthKit>=0.4.0
FormEncode>=0.6
pylibravatar
pylons
pytz
vobject
